Refactoring of create_events.py
